requirements.txt
git+https://github.com/LyzrCore/lyzr-automata.git@main
python-dotenv==1.0.1
streamlit==1.33.0